% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{apply_workflow_chain}
\alias{apply_workflow_chain}
\title{Apply workflow chain}
\usage{
apply_workflow_chain(chain_name, file_set, stop_on_error = TRUE)
}
\arguments{
\item{chain_name}{Name of workflow chain}

\item{file_set}{File set to process}

\item{stop_on_error}{Stop if any workflow fails}
}
\value{
Results from all workflows in chain
}
\description{
Apply workflow chain
}
