% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{convert_to_workflow}
\alias{convert_to_workflow}
\title{Generate workflow from existing analysis code}
\usage{
convert_to_workflow(input_line = NULL, output_lines = NULL)
}
\arguments{
\item{input_line}{Line that reads the data (to identify what to replace)}

\item{output_lines}{Lines that save outputs (to identify what to parameterize)}
}
\description{
Helper to convert existing analysis into workflow format
}
