% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{create_bibliography_template}
\alias{create_bibliography_template}
\title{Create bibliography template}
\usage{
create_bibliography_template(project_code = NULL)
}
\arguments{
\item{project_code}{Project code (optional, for project-specific bib)}
}
\description{
Sets up basic bibliography file in personal_notes
}
