% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{create_chain_from_master}
\alias{create_chain_from_master}
\title{Create workflow chain from master script analysis}
\usage{
create_chain_from_master(project_code, chain_name = NULL)
}
\arguments{
\item{project_code}{Project code to analyze}

\item{chain_name}{Name for the workflow chain}
}
\description{
Create workflow chain from master script analysis
}
