% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_collaborator_onboarding}
\alias{create_collaborator_onboarding}
\title{Create collaborator onboarding package}
\usage{
create_collaborator_onboarding(
  project_code = NULL,
  collaborator_name = "New Collaborator",
  package_type = "mixed",
  focus_areas = "all"
)
}
\arguments{
\item{project_code}{Project code}

\item{collaborator_name}{Name of collaborator being onboarded}

\item{package_type}{Type of package: "executive", "tutorial", "operational", "mixed"}

\item{focus_areas}{Areas of focus: "analysis", "writing", "methods", "domain", "all"}
}
\value{
Path to created onboarding package
}
\description{
Generates information package to bring new collaborators up to speed
}
