% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_executive_summary_doc}
\alias{create_executive_summary_doc}
\title{Create executive summary document}
\usage{
create_executive_summary_doc(project_code, output_dir)
}
\arguments{
\item{project_code}{Project code}

\item{output_dir}{Directory to save document}
}
\description{
High-level strategic overview for the project
}
