% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{create_folder_marker}
\alias{create_folder_marker}
\title{Create folder marker}
\usage{
create_folder_marker(folder_path, folder_name, folder_id, project_code)
}
\arguments{
\item{folder_path}{Path to folder}

\item{folder_name}{Folder name}

\item{folder_id}{Folder ID}

\item{project_code}{Project code}
}
\description{
Create folder marker
}
\keyword{internal}
