% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_operational_details_doc}
\alias{create_operational_details_doc}
\title{Create operational details document}
\usage{
create_operational_details_doc(project_code, output_dir)
}
\arguments{
\item{project_code}{Project code}

\item{output_dir}{Directory to save document}
}
\description{
Technical implementation and hands-on information
}
