% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_tutorial_walkthrough_doc}
\alias{create_tutorial_walkthrough_doc}
\title{Create tutorial walkthrough document}
\usage{
create_tutorial_walkthrough_doc(project_code, output_dir)
}
\arguments{
\item{project_code}{Project code}

\item{output_dir}{Directory to save document}
}
\description{
Educational explanation of research process and methods
}
