% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_update_template}
\alias{create_update_template}
\title{Create project update email template}
\usage{
create_update_template(project_code = NULL, update_type = "progress")
}
\arguments{
\item{project_code}{Project code}

\item{update_type}{Type of update: "progress", "milestone", "request", "results"}
}
\value{
Path to email template file
}
\description{
Generates template for regular project updates to collaborators
}
