% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_logging.R
\name{detect_file_sets}
\alias{detect_file_sets}
\title{Detect file sets by pattern matching}
\usage{
detect_file_sets(files)
}
\arguments{
\item{files}{Vector of filenames}
}
\value{
Named list of file sets
}
\description{
Internal function to group files by common naming patterns
}
