% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{detect_output_mode}
\alias{detect_output_mode}
\title{Detect output mode from workflow code}
\usage{
detect_output_mode(code_lines)
}
\arguments{
\item{code_lines}{Lines of workflow code}
}
\value{
"single" or "file_set"
}
\description{
Detect output mode from workflow code
}
