% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_logging.R
\name{detect_prefix_sets}
\alias{detect_prefix_sets}
\title{Detect sets by common filename prefixes}
\usage{
detect_prefix_sets(files)
}
\arguments{
\item{files}{Vector of filenames}
}
\value{
Named list of prefix-based sets
}
\description{
Groups remaining files by shared prefixes
}
