% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{detect_project_file_sets}
\alias{detect_project_file_sets}
\title{Detect file sets in current project}
\usage{
detect_project_file_sets(project_path)
}
\arguments{
\item{project_path}{Path to project}
}
\value{
Named list of file sets
}
\description{
Automatically identifies file sets for workflow configuration
}
