% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{detect_workflow_io}
\alias{detect_workflow_io}
\title{Detect inputs and outputs in workflow code}
\usage{
detect_workflow_io(code_lines)
}
\arguments{
\item{code_lines}{Lines of workflow code}
}
\value{
List with I/O information
}
\description{
Detect inputs and outputs in workflow code
}
