% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{export_workflow}
\alias{export_workflow}
\title{Export workflow for sharing}
\usage{
export_workflow(workflow_name, output_file = NULL)
}
\arguments{
\item{workflow_name}{Workflow to export}

\item{output_file}{Where to save (defaults to workflow_name.R)}
}
\description{
Create standalone version of workflow
}
