% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{extract_stage2_info}
\alias{extract_stage2_info}
\title{Extract Stage II information from project}
\usage{
extract_stage2_info(project_code, paths)
}
\arguments{
\item{project_code}{Project code}

\item{paths}{Project paths}
}
\value{
List of extracted information
}
\description{
Pulls information from analyseblitz files if they exist
}
