% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{extract_stage3_info}
\alias{extract_stage3_info}
\title{Extract Stage III information from project}
\usage{
extract_stage3_info(project_code, paths)
}
\arguments{
\item{project_code}{Project code}

\item{paths}{Project paths}
}
\value{
List of extracted information
}
\description{
Pulls information from draft documents if they exist
}
