% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback.R
\name{fallback}
\alias{fallback}
\title{Restore files from fallback checkpoint}
\usage{
fallback(path = ".")
}
\arguments{
\item{path}{Project directory (default: current directory)}
}
\description{
Restore files from fallback checkpoint
}
