% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{flag_suspicious_values}
\alias{flag_suspicious_values}
\title{Flag suspicious values in dataset}
\usage{
flag_suspicious_values(data, suspicious_codes = c(-999, -99, -9, 999, 9999))
}
\arguments{
\item{data}{Data frame to check}

\item{suspicious_codes}{Vector of codes to flag as suspicious}
}
\value{
Data frame of flagged values
}
\description{
Identifies potentially problematic values that may represent missing data codes
}
