% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{generate_component_template}
\alias{generate_component_template}
\title{Generate component template based on component type}
\usage{
generate_component_template(component, project_code, file_sets)
}
\arguments{
\item{component}{Component name}

\item{project_code}{Project code}

\item{file_sets}{Available file sets}
}
\value{
Character vector with component template
}
\description{
Creates appropriate template for each component type
}
