% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{generate_simulation_code}
\alias{generate_simulation_code}
\title{Generate data simulation code based on Stage I specs}
\usage{
generate_simulation_code(
  project_code,
  sample_size = 100,
  variables = NULL,
  analysis_type = "general"
)
}
\arguments{
\item{project_code}{Project code}

\item{sample_size}{Sample size from Stage I card}

\item{variables}{List of key variables from Stage I card}

\item{analysis_type}{Type of analysis planned}
}
\value{
Character string with simulation code
}
\description{
Creates R code for data simulation matching Stage I card specifications
}
