% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_file_sets}
\alias{get_file_sets}
\title{Get file sets as data structure}
\usage{
get_file_sets(folder_path = NULL)
}
\arguments{
\item{folder_path}{Folder to analyze}
}
\description{
Returns file sets as a list for programmatic use
}
