% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_files_in_set}
\alias{get_files_in_set}
\title{Get files in a file set}
\usage{
get_files_in_set(set_name)
}
\arguments{
\item{set_name}{Name of file set}
}
\value{
Vector of file names
}
\description{
Get files in a file set
}
