% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{init_blitz_project}
\alias{init_blitz_project}
\title{Initialize a project for Blitzschreiben workflow}
\usage{
init_blitz_project(project_code, idea_title)
}
\arguments{
\item{project_code}{Project code (e.g., "P3.1")}

\item{idea_title}{Title of the research idea}
}
\description{
Adds Blitzschreiben tracking to an existing PRoMan project
}
