% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{null_coalesce}
\alias{null_coalesce}
\title{Null-coalescing operator}
\usage{
null_coalesce(x, y)
}
\arguments{
\item{x}{First value to check}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null-coalescing operator
}
\keyword{internal}
