% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{pearl_growing_tracker}
\alias{pearl_growing_tracker}
\title{Pearl growing search tracker}
\usage{
pearl_growing_tracker(seed_paper, direction = "upstream", found_count = 0)
}
\arguments{
\item{seed_paper}{Main paper for pearl growing}

\item{direction}{"upstream" (references) or "downstream" (citing papers)}

\item{found_count}{Number of relevant papers found}
}
\description{
Helps track upstream/downstream reference searches
}
