% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{read_stage2_info}
\alias{read_stage2_info}
\title{Read Stage II analyseblitz information}
\usage{
read_stage2_info(project_code)
}
\arguments{
\item{project_code}{Project code}
}
\value{
List with methods and results info
}
\description{
Extracts key results and methods from Stage II template
}
