% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{remove_workflow}
\alias{remove_workflow}
\title{Remove a workflow}
\usage{
remove_workflow(workflow_name, confirm = TRUE)
}
\arguments{
\item{workflow_name}{Name of workflow to remove}

\item{confirm}{Ask for confirmation}
}
\description{
Remove a workflow
}
