% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{select_recent_files}
\alias{select_recent_files}
\title{Select recently modified files}
\usage{
select_recent_files(minutes = 30, folder_path = NULL)
}
\arguments{
\item{minutes}{How many minutes back to look (default: 30)}

\item{folder_path}{Folder to check (default: data/source)}
}
\value{
Vector of recently modified filenames
}
\description{
Quick function to get files modified in last N minutes
}
