% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{stage2_progress_check}
\alias{stage2_progress_check}
\title{Stage II progress tracker}
\usage{
stage2_progress_check(project_code = NULL)
}
\arguments{
\item{project_code}{Project code (optional, checks all if NULL)}
}
\description{
Shows current status of Stage II analyseblitz
}
