% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{test_stage1_simulation}
\alias{test_stage1_simulation}
\title{Test data simulation from Stage I}
\usage{
test_stage1_simulation(project_code, n = NULL)
}
\arguments{
\item{project_code}{Project code}

\item{n}{Sample size (overrides Stage I if provided)}
}
\description{
Generates and displays sample simulated data based on Stage I specs
}
