% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_logging.R
\name{track_project_access}
\alias{track_project_access}
\title{Track project access for session logging}
\usage{
track_project_access(project_path)
}
\arguments{
\item{project_path}{Current project path}
}
\description{
Internal function called when project is accessed
}
