% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{update_tracker_entry}
\alias{update_tracker_entry}
\title{Remove or update project in Blitzschreiben tracker}
\usage{
update_tracker_entry(project_code, action = "remove")
}
\arguments{
\item{project_code}{Project code to remove/update}

\item{action}{"remove" to delete, "archive" to mark as archived, "pause" to pause}
}
\description{
Removes a project from tracking or changes its status
}
