% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{verify_folder_marker}
\alias{verify_folder_marker}
\title{Verify folder marker}
\usage{
verify_folder_marker(folder_path, folder_name, folder_id)
}
\arguments{
\item{folder_path}{Path to folder}

\item{folder_name}{Expected folder name}

\item{folder_id}{Expected folder ID}
}
\value{
Logical indicating if marker is valid
}
\description{
Verify folder marker
}
\keyword{internal}
