% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{write_proman}
\alias{write_proman}
\title{Write updates to .proman file}
\usage{
write_proman(proman_data, clear_cache = TRUE)
}
\arguments{
\item{proman_data}{The modified proman data to write}

\item{clear_cache}{Whether to clear the cache after writing}
}
\description{
Handles the complexity of updating hidden files on Windows
}
\keyword{internal}
